+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `GeneralMicrosimulation` alias for `Microsimulation`.
 
+### Fixed
+
+* `is_in` function takes list arguments and works correctly.
+
 ## [0.2.3] - 2022-01-26
 
 ### Fixed